add future call

+++ lyon/streams/socket_stream.cc
@@ -4,20 +4,20 @@ namespace lyon {
 
 static Logger::ptr g_logger = LYON_LOG_GET_LOGGER("system");
 
-int SocketStream::read(void *buffer, size_t size) {
+ssize_t SocketStream::read(void *buffer, size_t size) {
     if (isConnected())
         return m_socket->recv(buffer, size);
     return -1;
 }
 
-int SocketStream::read(ByteArray::ptr bytearray, size_t size) {
+ssize_t SocketStream::read(ByteArray::ptr bytearray, size_t size) {
     if (!isConnected()) {
         return -1;
     }
     std::vector<iovec> buffers;
     bytearray->getWriteBuffer(buffers, size);
     // LYON_LOG_DEBUG(g_logger) << "start sock read";
-    int rt = m_socket->recv(&buffers[0], buffers.size());
+    ssize_t rt = m_socket->recv(&buffers[0], buffers.size());
     // LYON_LOG_DEBUG(g_logger) << "sock read rt = " << rt;
     if (rt > 0) {
         bytearray->setPosition(bytearray->getPosition() + rt);
@@ -25,20 +25,20 @@ int SocketStream::read(ByteArray::ptr bytearray, size_t size) {
     return rt;
 }
 
-int SocketStream::write(const void *buffer, size_t size) {
+ssize_t SocketStream::write(const void *buffer, size_t size) {
     if (isConnected()) {
         return m_socket->send(buffer, size);
     }
     return -1;
 }
 
-int SocketStream::write(ByteArray::ptr bytearray, size_t size) {
+ssize_t SocketStream::write(ByteArray::ptr bytearray, size_t size) {
     if (!isConnected()) {
         return -1;
     }
     std::vector<iovec> buffers;
     bytearray->getReadBuffer(buffers, size);
-    int rt = m_socket->send(&buffers[0], buffers.size());
+    ssize_t rt = m_socket->send(&buffers[0], buffers.size());
     if (rt > 0) {
         bytearray->setPosition(bytearray->getPosition() + rt);
     }
